+++ miso-dto/src/main/java/uk/ac/bbsrc/tgac/miso/dto/ArrayModelDto.java
@@ -4,6 +4,8 @@ public class ArrayModelDto {
 
   private Long id;
   private String alias;
+  private Integer rows;
+  private Integer Columns;
 
   public Long getId() {
     return id;
@@ -19,6 +21,22 @@ public class ArrayModelDto {
 
   public void setAlias(String alias) {
     this.alias = alias;
+  }
+
+  public Integer getRows() {
+    return rows;
+  }
+
+  public void setRows(Integer rows) {
+    this.rows = rows;
+  }
+
+  public Integer getColumns() {
+    return Columns;
+  }
+
+  public void setColumns(Integer columns) {
+    Columns = columns;
   }
 
 }